switch to require_modules/patch setup
